# real-time-notifications-microservice/src/main/resources/static/openapi.yaml
openapi: 3.0.1
info:
  title: Real Time Notifications System
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  /api-clients/v1.0/notifications/users/{userId}:
    get:
      tags:
      - Notifications
      summary: Get user notifications
      description: Retrieve a list of notifications for a specific user
      operationId: getUserNotifications
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successfully retrieved notifications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomApiResponseListNotification"
    post:
      tags:
      - Notifications
      summary: Create a notification for a user
      description: Create a new notification for a specific user
      operationId: createNotification
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Notification"
        required: true
      responses:
        "201":
          description: Notification successfully created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
        "400":
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomApiResponseNotification"
  /api-clients/v1.0/notifications/stream/{userId}:
    get:
      tags:
      - Notifications
      summary: Stream user notifications
      description: Stream notifications for a specific user in real time
      operationId: streamNotifications
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Streaming notifications
          content:
            text/event-stream:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    Notification:
      type: object
      properties:
        id:
          type: string
        userReferenceId:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
    CustomApiResponseNotification:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
        data:
          $ref: "#/components/schemas/Notification"
...
    CustomApiResponseListNotification:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/Notification"
...